+++ src/Lab1/interfaces/IMovable.java
@@ -1,4 +1,4 @@
-package Lab1;
+package Lab1.interfaces;
 
 public interface IMovable {
     public void move();
