+++ src/Lab1/truckextensions/Cartransporter.java
@@ -1,4 +1,6 @@
-package Lab1;
+package Lab1.truckextensions;
+
+import Lab1.interfaces.ITrailer;
 
 public class Cartransporter implements ITrailer {
     private Boolean extendedRamp;
